requirements.txt
langchain>=0.1.0
langchain-community>=0.0.20
langchain-groq>=0.1.0
langchain-text-splitters>=0.0.1
langchain-huggingface>=0.0.1
faiss-cpu>=1.7.4
pypdf>=3.0.0
streamlit>=1.28.0
python-dotenv>=1.0.0
sentence-transformers>=2.2.0
torch==2.3.0+cpu
transformers>=4.30.0